feat: implement arch unit tests (#2)
* feat: implement arch unit tests

* feat: using bigdecimal type

* refactor: implement service domain

* Refactor core layer (#3)

* refactor: some code adjust

* feat: full payment flow

* feat: show balance

+++ src/main/java/com/payments/api/core/entities/payments/Payee.java
@@ -1,7 +1,9 @@
 package com.payments.api.core.entities.payments;
 
+import java.math.BigDecimal;
+
 public interface Payee {
 
-    void credit(double value);
+    void credit(BigDecimal value);
 
 }
